major cleanup of main, structural changes, viewing full tree in UI, cursor.Up() added
