+++ keyboard.go
@@ -1,0 +1,24 @@
+package main
+
+import (
+	"github.com/gdamore/tcell/v2"
+	"os"
+)
+
+func handleEventKey(ev *tcell.EventKey, s tcell.Screen, c *Cursor) {
+	switch ev.Key() {
+	case tcell.KeyEscape:
+		s.Fini()
+		os.Exit(0)
+	case tcell.KeyEnter:
+		newItem(c.i)
+	case tcell.KeyBackspace2:
+	case tcell.KeyUp:
+	case tcell.KeyDown:
+	case tcell.KeyRune:
+	case tcell.KeyRight:
+	case tcell.KeyLeft:
+	case tcell.KeyTab:
+	case tcell.KeyBacktab:
+	}
+}
